requirements.txt
slack_sdk==3.34.0
python-dotenv==1.0.1
jpholiday==0.1.10
